+++ src/render_init.cpp
@@ -142,7 +142,7 @@ SetDrawBuffers(framebuffer *FBO)
     Attachments[AttIndex] =  GL_COLOR_ATTACHMENT0 + AttIndex;
   }
 
-  glDrawBuffers(FBO->Attachments, Attachments);
+  glDrawBuffers((s32)FBO->Attachments, Attachments);
 }
 
 shader
@@ -271,7 +271,7 @@ InitializeShadowBuffer(shadow_render_group *SG, memory_arena *GraphicsMemory, v2
   SG->DebugTextureShader = MakeSimpleTextureShader(SG->ShadowMap, GraphicsMemory);
 
   SG->DepthShader = LoadShaders( "DepthRTT.vertexshader", "DepthRTT.fragmentshader", GraphicsMemory);
-  SG->MVP_ID = GetShaderUniform(&SG->DepthShader, "depthMVP");
+  SG->MVP_ID = (u32)GetShaderUniform(&SG->DepthShader, "depthMVP");
 
   if(glCheckFramebufferStatus(GL_FRAMEBUFFER) != GL_FRAMEBUFFER_COMPLETE)
     return false;
